backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
pytest==7.4.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
